Replace platform-dependent long with fixed-width integer types (#773)
* Give the millisecond clock and nVars() fixed-width types

"long" is 64 bits on LP64 but 32 on ILP32 and LLP64, so these two
interfaces meant different things depending on the target.

getCurrentTime() returned a millisecond count in a long. Where time_t
is 32 bits -- the default on i386 -- "1000 * t.tv_sec" overflowed
before the return type was reached, which is undefined, and the result
did not fit a 32-bit long anyway. It survived only because every caller
takes differences, and the wraparound cancels for short intervals.

Widen tv_sec before the multiply and return int64_t. RunTimes' Element,
times map, addTime() and lastTime follow, as do print()'s millisecond
accumulators, which were int while summing 64-bit values.

The function had two byte-identical definitions, in RunTimes.cpp and
ASTmisc.cpp. Keep the one in Util, where the other timing code lives,
and move the declaration to RunTimes.h. No include changes are needed:
AST.h already reaches RunTimes.h through UsefulDefs.h.

nVars() returned unsigned long from every solver. All five underlying
solvers return int or uint32_t, and the one caller assigns straight
back to an int, so the width was 64 bits on Linux, 32 on i386 and
Win64, and meaningful nowhere. Make it uint32_t.

rewrite_rule_gen's RewriteRule::setVerified() takes an int, so the
elapsed durations passed to it get an explicit cast rather than an
implicit narrowing.

* Remove the remaining word-size dependencies

None of these were wrong on LP64, but each meant something different
where long is 32 bits.

CreateBVConst chopped its uint64 argument into chunks the width of a
machine word. That happened to be correct on both word sizes, but only
because Chunk_Store also takes an unsigned long, and it needed a
-Wshift-count-overflow suppression for a shift that is dead on 64-bit
and live on 32. Store 32 bits at a time instead, which is what CBVOps
already does for the same reason, and the coupling and the pragma both
go away.

Checked against an independent oracle over widths 1 to 130 -- either
side of both chunk boundaries -- for boundary values and for every
single-bit value: 10400 cases, all agreeing. Deliberately mis-chunking
the loop fails 3294 of them, so the check is not vacuous.

MersenneTwister typedef'd its state word to unsigned long, so the state
array was twice the necessary size on LP64 and the /dev/urandom seed
read 8 bytes per word to then discard 4. The algorithm masks every state
update to 32 bits -- initialize() and twist() both -- so the generated
sequence is unchanged; the seeded exhaustive tests confirm it.

test_fprewrites computed a symbol's domain as "1ul << width", which is
fine on LP64 and undefined where long is 32 bits, and only checked the
enumeration limit after the product was formed. Saturate the shift and
test the limit before each multiply, so neither can overflow.

The rest is spelling: loop counters and widths that were long or
unsigned long for no reason.

* Convert the last two assignment-enumeration loops to uint64_t

Same pattern as the three converted in the previous commit; these two
were hidden from the first survey by a comment filter.

* Take the buffer lengths in the C API as size_t

The four print-to-buffer entry points returned their length through an
"unsigned long*". That is 64 bits on Linux and macOS but 32 on i386 and
on 64-bit Windows, so the one installed header described a different
function depending on the target.

size_t is what these actually are: every implementation stores
s.size() + 1 into them. It is also the least disruptive of the
consistent choices -- identical to the current type on Linux and macOS
at both word sizes, and on i386. Only 64-bit Windows callers see a
width change, from 32 to 64 bits.

This is an ABI change to installed API, so it needs a version note.

The ctypes bindings declared the same parameter as c_ulong, which had
the matching bug; they now use c_size_t. The malloc failure messages
move from %lu to %zu.

* Spell the CreateBVConst value parameter uint64_t

The three CreateBVConst overloads that take a value took it as an
"unsigned long long int". That is 64 bits everywhere, so unlike the
other conversions in this series it was never a portability bug -- it
is a spelling change, for consistency with the rest of the interface.

It is not a no-op, though. On LP64 uint64_t is unsigned long, a type
distinct from unsigned long long despite the identical width, so the
mangled names change:

  _ZN3stp6STPMgr13CreateBVConstEjy  ->  _ZN3stp6STPMgr13CreateBVConstEjm

and likewise for Cpp_interface and NodeFactory. That is an ABI break on
Linux and macOS. On i386 and Windows, where long is 32 bits, uint64_t
already is unsigned long long and the symbols are unchanged.

Taken deliberately: the next release carries other ABI breaks.

The headers now include <cstdint> rather than relying on picking it up
transitively, and test_fprewrites drops a cast that is now redundant.

+++ include/stp/NodeFactory/NodeFactory.h
@@ -27,6 +27,7 @@ THE SOFTWARE.
 
 #include "stp/AST/ASTKind.h"
 #include "stp/Util/Attributes.h"
+#include <cstdint>
 #include <vector>
 
 using std::vector;
@@ -107,7 +108,7 @@ public:
   ASTNode CreateMaxConst(unsigned width);
   ASTNode CreateSignedMinConst(unsigned width);
   
-  ASTNode CreateBVConst(unsigned int width, unsigned long long int bvconst);
+  ASTNode CreateBVConst(unsigned int width, uint64_t bvconst);
   ASTNode CreateFPConst(const stp::ASTNode& bvconst, unsigned exp_width,
                         unsigned sig_width);
